DbConnectionBroker.java is now a wrapper for the c3p0 database connection broker.
A lot of stuff out of it wasn't used, so was removed instead of ported. Some other things remain, unused, marked as deprecated.

Many more exceptions are now thrown, instead of nulls returned. All of the current calls to getConnection() fail reasonably sanely.

freeConnection() should possibly be deprecated, but I left it just in case.

This will hopefully fix #7 and #11.