Seen and Tell hardened against spamming the channel when mysql goes down., neither now produce error messages (they ignore their instead job) if an sqlexception occurs.
This behaviour can be overridden in Tell using the tell.get (which will try again and give an error), seen isn't really that critical.

In this light, Seen will lockdown for 2 minutes if an sql exception occurs in it, and will return instantly (hence not requiring db access), and not unnecessarily blocking threads, in theory anything that doesn't require db access (not much, thanks to aliases) can continue to be used.

c3p0 seems to cope resonably well with the loss/regain of mysql, occasionally on it's necessary to use up all of the dud connections, not much we can do about this. Resetting the connection pool would invalidate any that are working, test on checkout has already been ruled out on speed issues.

Oh, and a rollup of a change to miscmsg's random that was missed before.