Updated the entirety of SecurityModule.java to take events (UserEvents) instead of a String for the nick, allowing actual checking to be done on the event. This addresses #143.
The current check on the event is that it wasn't created under (an arbitrary) 5 seconds ago. Failing this, it'll throw a ChoobEventExpired exception. No extra information has been embedded in this as nick/user/hostname are pointless.

This is a minor hack, but allows aliases to continue working. Synthetic level checking may be worth it.. maybe attaching the user's "security level" (by nodeid?) to the event when it's sent. Ticket with ideas.

It turns out that event validation only appears to be necessary in two places; getNodeIDFromUserName and hasNS. checkNS, checkNickPerm, checkPerm, hasPerm, hasNickPerm and bindPlugin are fully dependant on these.

add*, getRootUser, linkUser and removeUserFromGroup are relatively high risk commands that can't be checked in this way. Extra care may need to be taken.

A few places still throw a possibly erroneous ChoobException("User " + userEvent.getNick() + ".., this shouldn't be reached if the event has a rogue nick in it.

Tested with core plugins, needs testing with all other plugins, should only effect the ones that use SecurityModule.java.

And, of course, the obligatory javadoc regen.