Change to queue certain messages to aid in avoiding flooding out. Queues and potentially delays private messages while allowing channel messages to be sent without hindrance.
There are three queues, one queue is the outbound queue, that gets sent directly to the server. The second queue is the channel queue (target starts with # - there's probably a better test for this...) which has no significant delay, and just gets added straight onto the outbound queue. The third queue is private messages, which, if the size is > 3 then the messages are added one every 3 seconds to the outbound queue.